+++ approval-server/src/main/java/fun/sast/service/CompetitionService.java
@@ -3,5 +3,5 @@ package fun.sast.service;
 import com.alibaba.fastjson2.JSONObject;
 
 public interface CompetitionService {
-    JSONObject getSchema(Long comId);
+    JSONObject getSchema(Integer comId);
 }
